InMemoryAssemblyProvider no longer hard coded loader
