+++ src/Compilers/Core/CodeAnalysisTest/AnalyzerFileReferenceTests.cs
@@ -15,6 +15,11 @@ namespace Microsoft.CodeAnalysis.UnitTests
 {
     public class AnalyzerFileReferenceTests : TestBase
     {
+        public static AnalyzerFileReference CreateAnalyzerFileReference(string fullPath)
+        {
+            return new AnalyzerFileReference(fullPath, InMemoryAssemblyProvider.GetAssembly);
+        }
+
         [Fact]
         public void AssemblyLoading()
         {
@@ -28,14 +33,14 @@ namespace Microsoft.CodeAnalysis.UnitTests
                 sb.AppendLine();
             };
 
-            AnalyzerFileReference.AssemblyLoad += handler;
+            InMemoryAssemblyProvider.AssemblyLoad += handler;
 
             var alphaDll = directory.CreateFile("Alpha.dll").WriteAllBytes(TestResources.AssemblyLoadTests.AssemblyLoadTests.Alpha);
             var betaDll = directory.CreateFile("Beta.dll").WriteAllBytes(TestResources.AssemblyLoadTests.AssemblyLoadTests.Beta);
             var gammaDll = directory.CreateFile("Gamma.dll").WriteAllBytes(TestResources.AssemblyLoadTests.AssemblyLoadTests.Gamma);
             var deltaDll = directory.CreateFile("Delta.dll").WriteAllBytes(TestResources.AssemblyLoadTests.AssemblyLoadTests.Delta);
 
-            AnalyzerFileReference alphaReference = new AnalyzerFileReference(alphaDll.Path);
+            AnalyzerFileReference alphaReference = CreateAnalyzerFileReference(alphaDll.Path);
             Assembly alpha = alphaReference.GetAssembly();
             File.Delete(alphaDll.Path);
 
@@ -45,7 +50,7 @@ namespace Microsoft.CodeAnalysis.UnitTests
             File.Delete(gammaDll.Path);
             File.Delete(deltaDll.Path);
 
-            AnalyzerFileReference betaReference = new AnalyzerFileReference(betaDll.Path);
+            AnalyzerFileReference betaReference = CreateAnalyzerFileReference(betaDll.Path);
             Assembly beta = betaReference.GetAssembly();
             var b = beta.CreateInstance("Beta.B");
             b.GetType().GetMethod("Write").Invoke(b, new object[] { sb, "Test B" });
@@ -62,23 +67,23 @@ Delta: Gamma: Beta: Test B
 
             Assert.Equal(expected, actual);
 
-            var alphaDllRequestor = AnalyzerFileReference.TryGetRequestingAssemblyPath(alphaDll.Path);
-            var betaDllRequestor = AnalyzerFileReference.TryGetRequestingAssemblyPath(betaDll.Path);
-            var gammaDllRequestor = AnalyzerFileReference.TryGetRequestingAssemblyPath(gammaDll.Path);
-            var deltaDllRequestor = AnalyzerFileReference.TryGetRequestingAssemblyPath(deltaDll.Path);
+            var alphaDllRequestor = InMemoryAssemblyProvider.TryGetRequestingAssembly(alphaDll.Path);
+            var betaDllRequestor = InMemoryAssemblyProvider.TryGetRequestingAssembly(betaDll.Path);
+            var gammaDllRequestor = InMemoryAssemblyProvider.TryGetRequestingAssembly(gammaDll.Path);
+            var deltaDllRequestor = InMemoryAssemblyProvider.TryGetRequestingAssembly(deltaDll.Path);
 
             Assert.Null(alphaDllRequestor);
             Assert.Null(betaDllRequestor);
             Assert.Equal(expected: alphaDll.Path, actual: gammaDllRequestor, comparer: StringComparer.OrdinalIgnoreCase);
             Assert.Equal(expected: gammaDll.Path, actual: deltaDllRequestor, comparer: StringComparer.OrdinalIgnoreCase);
 
-            AnalyzerFileReference.AssemblyLoad -= handler;
+            InMemoryAssemblyProvider.AssemblyLoad -= handler;
         }
 
         [Fact]
         public void TestMetadataParse()
         {
-            AnalyzerFileReference reference = new AnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
             var analyzerTypeNameMap = reference.GetAnalyzerTypeNameMap();
             Assert.Equal(2, analyzerTypeNameMap.Keys.Count());
 
@@ -104,7 +109,7 @@ Delta: Gamma: Beta: Test B
         [Fact]
         public void TestGetAnalyzersPerLanguage()
         {
-            AnalyzerFileReference reference = new AnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
             var analyzers = reference.GetAnalyzers(LanguageNames.CSharp);
             Assert.Equal(4, analyzers.Length);
             var analyzerNames = analyzers.Select(a => a.GetType().Name);
@@ -125,7 +130,7 @@ Delta: Gamma: Beta: Test B
         [Fact]
         public void TestLoadErrors1()
         {
-            AnalyzerFileReference reference = new AnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(Assembly.GetExecutingAssembly().Location);
 
             List<AnalyzerLoadFailureEventArgs> errors = new List<AnalyzerLoadFailureEventArgs>();
             EventHandler<AnalyzerLoadFailureEventArgs> errorHandler = (o, e) => errors.Add(e);
@@ -144,7 +149,7 @@ Delta: Gamma: Beta: Test B
         [Fact]
         public void TestLoadErrors2()
         {
-            AnalyzerFileReference reference = new AnalyzerFileReference("C:\\randomlocation\\random.dll");
+            AnalyzerFileReference reference = CreateAnalyzerFileReference("C:\\randomlocation\\random.dll");
 
             List<AnalyzerLoadFailureEventArgs> errors = new List<AnalyzerLoadFailureEventArgs>();
             EventHandler<AnalyzerLoadFailureEventArgs> errorHandler = (o, e) => errors.Add(e);
@@ -163,7 +168,7 @@ Delta: Gamma: Beta: Test B
         {
             var directory = Temp.CreateDirectory();
             var alphaDll = directory.CreateFile("Alpha.dll").WriteAllBytes(TestResources.AssemblyLoadTests.AssemblyLoadTests.Alpha);
-            AnalyzerFileReference reference = new AnalyzerFileReference(alphaDll.Path);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(alphaDll.Path);
 
             List<AnalyzerLoadFailureEventArgs> errors = new List<AnalyzerLoadFailureEventArgs>();
             EventHandler<AnalyzerLoadFailureEventArgs> errorHandler = (o, e) => errors.Add(e);
@@ -185,7 +190,7 @@ Delta: Gamma: Beta: Test B
         {
             var directory = Temp.CreateDirectory();
             var textFile = directory.CreateFile("Foo.txt").WriteAllText("I am the very model of a modern major general.");
-            AnalyzerFileReference reference = new AnalyzerFileReference(textFile.Path);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(textFile.Path);
 
             Assert.Equal(expected: "Foo.txt", actual: reference.Display);
         }
@@ -196,7 +201,7 @@ Delta: Gamma: Beta: Test B
         {
             var directory = Temp.CreateDirectory();
             var analyzerDll = directory.CreateFile("Alpha.dll").WriteAllBytes(TestResources.AnalyzerTests.AnalyzerTests.FaultyAnalyzer);
-            AnalyzerFileReference reference = new AnalyzerFileReference(analyzerDll.Path);
+            AnalyzerFileReference reference = CreateAnalyzerFileReference(analyzerDll.Path);
 
             List<AnalyzerLoadFailureEventArgs> errors = new List<AnalyzerLoadFailureEventArgs>();
             EventHandler<AnalyzerLoadFailureEventArgs> errorHandler = (o, e) => errors.Add(e);
