+++ src/EditorFeatures/Test2/Diagnostics/DiagnosticServiceTests.vb
@@ -19,6 +19,10 @@ Namespace Microsoft.CodeAnalysis.Editor.Implementation.Diagnostics.UnitTests
 
     Public Class DiagnosticServiceTests
 
+        Public Function CreateAnalyzerFileReference(ByVal fullPath As String) As AnalyzerFileReference
+            Return New AnalyzerFileReference(fullPath, AddressOf InMemoryAssemblyProvider.GetAssembly)
+        End Function
+
         <Fact, Trait(Traits.Feature, Traits.Features.Diagnostics)>
         Public Sub TestProjectAnalyzers()
             Dim test = <Workspace>
@@ -302,8 +306,8 @@ Namespace Microsoft.CodeAnalysis.Editor.Implementation.Diagnostics.UnitTests
 
             Using workspace = TestWorkspaceFactory.CreateWorkspace(test)
                 Dim project = workspace.CurrentSolution.Projects.Single()
-                Dim analyzerReference1 = New AnalyzerFileReference("x:\temp.dll")
-                Dim analyzerReference2 = New AnalyzerFileReference("x:\temp.dll")
+                Dim analyzerReference1 = CreateAnalyzerFileReference("x:\temp.dll")
+                Dim analyzerReference2 = CreateAnalyzerFileReference("x:\temp.dll")
                 project = project.AddAnalyzerReference(analyzerReference1)
 #If DEBUG Then
                 Assert.Throws(Of TraceAssertException)(Function() project.AddAnalyzerReference(analyzerReference2))
@@ -324,10 +328,10 @@ Namespace Microsoft.CodeAnalysis.Editor.Implementation.Diagnostics.UnitTests
             Using workspace = TestWorkspaceFactory.CreateWorkspace(test)
                 ' Add duplicate analyzer references: one as VSIX analyzer reference and other one as project analyzer reference.
                 Dim project = workspace.CurrentSolution.Projects.Single()
-                Dim analyzerReference1 = New AnalyzerFileReference(Assembly.GetExecutingAssembly().Location)
+                Dim analyzerReference1 = CreateAnalyzerFileReference(Assembly.GetExecutingAssembly().Location)
                 project = project.AddAnalyzerReference(analyzerReference1)
 
-                Dim analyzerReference2 = New AnalyzerFileReference(Assembly.GetExecutingAssembly().Location)
+                Dim analyzerReference2 = CreateAnalyzerFileReference(Assembly.GetExecutingAssembly().Location)
                 Dim diagnosticService = New TestDiagnosticAnalyzerService(ImmutableArray.Create(Of AnalyzerReference)(analyzerReference2))
 
                 Dim analyzer = diagnosticService.CreateIncrementalAnalyzer(workspace)
